+++ ws2811.h
@@ -65,7 +65,7 @@ extern "C" {
 
 struct ws2811_device;
 
-typedef uint32_t ws2811_led_t;                   //< 0xWWRRGGBB
+typedef uint64_t ws2811_led_t;                   //< 0xWWRRGGBB
 typedef struct
 {
     int gpionum;                                 //< GPIO Pin with PWM alternate function, 0 if unused
@@ -86,7 +86,7 @@ typedef struct
     uint64_t render_wait_time;                   //< time in µs before the next render can run
     struct ws2811_device *device;                //< Private data for driver use
     const rpi_hw_t *rpi_hw;                      //< RPI Hardware Information
-    uint32_t freq;                               //< Required output frequency
+    uint64_t freq;                               //< Required output frequency
     int dmanum;                                  //< DMA number _not_ already in use
     ws2811_channel_t channel[RPI_PWM_CHANNELS];
 } ws2811_t;
